Move install code.
